:recycle: move impl FromRequest for User
